- Separate enums values in constants package
