+++ src/main/kotlin/com/piashcse/database/entities/Order.kt
@@ -1,5 +1,6 @@
 package com.piashcse.database.entities
 
+import com.piashcse.constants.OrderStatus
 import com.piashcse.database.entities.base.BaseIntEntity
 import com.piashcse.database.entities.base.BaseIntEntityClass
 import com.piashcse.database.entities.base.BaseIntIdTable
@@ -10,10 +11,6 @@ object OrderTable : BaseIntIdTable("order") {
     val subTotal = float("sub_total")
     val total = float("total")
     val status = enumerationByName("status", 30, OrderStatus::class).clientDefault { OrderStatus.PENDING }
-
-    enum class OrderStatus {
-        PENDING, CONFIRMED, PAID, DELIVERED, CANCELED, RECEIVED
-    }
 }
 
 class OrderDAO(id: EntityID<String>) : BaseIntEntity(id, OrderTable) {
@@ -35,5 +32,5 @@ data class Order(
     val orderId: String,
     val subTotal: Float,
     val total: Float,
-    val status: OrderTable.OrderStatus,
+    val status: OrderStatus,
 )
